image scanner ent changes (#6305)
