+++ pkg/policyGovernance/security/imageScanning/ImageScanService.go
@@ -22,7 +22,6 @@ import (
 	"github.com/devtron-labs/devtron/pkg/cluster/environment"
 	"github.com/devtron-labs/devtron/pkg/cluster/environment/bean"
 	bean2 "github.com/devtron-labs/devtron/pkg/deployment/trigger/devtronApps/bean"
-	"github.com/devtron-labs/devtron/pkg/policyGovernance/security/imageScanning/adapter"
 	bean3 "github.com/devtron-labs/devtron/pkg/policyGovernance/security/imageScanning/bean"
 	"github.com/devtron-labs/devtron/pkg/policyGovernance/security/imageScanning/helper/parser"
 	repository3 "github.com/devtron-labs/devtron/pkg/policyGovernance/security/imageScanning/repository"
@@ -712,22 +711,6 @@ func (impl ImageScanServiceImpl) IsImageScanExecutionCompleted(image, imageDiges
 		}
 	}
 	return isScanningCompleted, nil
-}
-
-func (impl ImageScanServiceImpl) GetScanResults(resourceScanQueryParams *bean3.ResourceScanQueryParams) (resp parser.ResourceScanResponseDto, err error) {
-	request := &bean3.ImageScanRequest{
-		ArtifactId: resourceScanQueryParams.ArtifactId,
-		AppId:      resourceScanQueryParams.AppId,
-		EnvId:      resourceScanQueryParams.EnvId,
-	}
-	respFromExecutionDetail, err := impl.FetchExecutionDetailResult(request)
-	if err != nil {
-		impl.Logger.Errorw("error encountered in GetScanResults", "req", request, "err", err)
-		return resp, err
-	}
-	// build an adapter to convert the respFromExecutionDetail to the required ResourceScanResponseDto format
-	return adapter.ExecutionDetailsToResourceScanResponseDto(respFromExecutionDetail), nil
-
 }
 
 func (impl ImageScanServiceImpl) FilterDeployInfoByScannedArtifactsDeployedInEnv(deployInfoList []*repository3.ImageScanDeployInfo) ([]*repository3.ImageScanDeployInfo, error) {
